[STM,FIX] Always shutdown USB EHCI before booting linux
Prior to this commit, and despite the name, the U-Boot command "usb
stop" did not actually stop the EHCI (USB 2.x) controller! In contrast,
"usb stop" did work correctly for OHCI (USB 1.x).

Within the confines of U-Boot, this is not a significant problem, as
U-Boot owns all the memory.  However, when chain-loading a subsequent
"payload application" (e.g. an OS), then this can be a very significant
problem indeed!  As a result, when linux was booted from U-Boot (after
starting the EHCI controller), then the EHCI controller would continue to
access (and modify) system RAM whilst linux was still initializing itself.
This clearly can result in dangerous and undefined behaviour!

This commit adds code to ensure that the EHCI controller is shutdown
cleanly when "usb stop" is performed. This new code will also be
*implicitly* called when "bootm" is used, if USB was started. Hence,
there is no need to call "usb stop" manually.

Tested on a single B2020 board with a STxH415 SoC, using the ST40 core.
Linux was not booted, instead the state of the EHCI (and OHCI) USB
controllers were dumped immediately prior to the point where control
would normally be passed to linux, and it was confirmed that both USB
controllers were not running.

Please see the following for further information:

	https://bugzilla.stlinux.com/show_bug.cgi?id=41055
	https://bugzilla.stlinux.com/show_bug.cgi?id=41767

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>