[STM] Add support for a linux-style "fixed PHY".
In preparation for adding support for the RealTek ethernet switch, as used
on the B2120 board (utilizing the STxH407 SoC) from STMicroelectronics,
we add support for a linux-style "fixed PHY" connected to the ST-MAC.

Using a similar convention to linux, we effectively treat the ST-MAC
as if it were connected to a "fixed PHY". This means that we do not
try and control it via writing to the traditional IEEE PHY register
set. Specifically, we do not try and "reset" it, nor effect the
auto-negotiation status of it, etc.

In addition, the user commands "mii read" and "mii write" will now just
issue an appropriate diagnostic informing the user that these commands
are not supported for this class of PHY device.

This new capability is enabled by defining the following macro:
	#define CONFIG_STMAC_USE_FIXED_PHY

Finally, the code will now refuse to build if both the following macros
are defined: CONFIG_STMAC_USE_FIXED_PHY and CONFIG_PHY_LOOPBACK. This
restriction could be removed in the future, but it is deemed to be an
acceptable limitation for the time-being.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>