[STM] Add support for RealTek RTL8367RB switch
In preparation for adding support for the RealTek switch, as used on the
B2120 board (utilizing the STxH407 SoC) from STMicroelectronics, we add
core support for the RealTek RTL8367RB ethernet switch, for the ST-GMAC.

It should be noted that this support enables the RealTek RTL8367RB switch
as a (dumb) "fixed PHY" only, and does not support the more advanced
features of this switch.

This new capability is enabled by defining the following macros:

	#define CONFIG_STMAC_RTL8367RB
	#define CONFIG_STMAC_USE_RTL8367_SMI

In addition, the following also need to be defined appropriately:

	#define CONFIG_STMAC_RTL8367_SMI_SDA <x>, <y>
	#define CONFIG_STMAC_RTL8367_SMI_SCK <x>, <y>

Even though the RealTek RTL8367RB switch is configured as a "fixed PHY",
the "mii read" command works (but not "mii write") for this particular
"fixed PHY". It should be possible to add support for "mii write",
but this is not seen as a strong requirement.

It is also possible to read, dump & reset the 64-bit MIB (statistics)
counters present on the RTL8367RB switch. This can be achieved by defining
the following macro:

	#define CONFIG_STM_RTL8367_MIB_SUPPORT

It should be noted that when the RealTek switch is initialized,
the "firmware" to be downloaded to it, is stored in the file
"drivers/stm/rtl8367_smi_fw.c". The firmware can be re-generated by
instrumenting linux, and post-processing the linux console output -
please see "drivers/stm/rtl8367_smi_fw.sed" for further details.

Finally, thanks to Gabor Juhos <juhosg@openwrt.org>, for providing the
original code for linux, on which parts of this commit are based.  Thanks!

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>