[STM] Original patches for eMMC/SD support for STMicroelectronics
This commit includes the original set of patches ("as is") for U-Boot,
to add core support for eMMC/SD using STMicroelectronics' MMC IP block,
as used on STMicroelectronics' STxH407 SoC, on the B2120 board.

It should be noted that this commit does not enable eMMC/SD support on
STMicroelectronics' systems, it only provides the basic infrastructure.

Thanks to Youssef Triki <Youssef.Triki@st.com> for the original patches.

Signed-off-by: Youssef Triki <Youssef.Triki@st.com>
Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>