[STM] rename struct stm_gmac_pin to stm_pad_pin
The "struct stm_gmac_pin" was named so, because it was used exclusively
to set up pins/pads specifically for the ST-GMAC. However, in preparation
for adding support to use the STMicroelectronics' MMC controller, which
will be using this structure, it makes sense to rename the struct to
remove the GMAC-specific nature, and make it more "generic". Hence, this
commit essentially renames "struct stm_gmac_pin" to "struct stm_pad_pin".

In addition, 3 of the fields in this structure are specific to the
GMAC (the remainder being "common") - these 3 fields are now placed in
a GMAC-specific structure inside a union in struct stm_pad_pin. This
just tidies up the structure making it clearer what is common, and what
is specific.

This commit does not change the code's logic – it is essentially just
a benign renaming exercise.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>