[STM] Re-work recent support for eMMC/SD for STMicroelectronics
Following on from the previous commit (which added the original patches
for eMMC/SD), this commit essentially re-works that recent commit,
after reviewing the patches. Specifically, this commit incorporates
the feedback from the aforementioned review, and makes the files more
consistent with the STMicroelectronics' tree. In addition, a logical
error in stxh407_mmc_getcd() was also corrected.

It should be noted that this commit does not enable eMMC/SD support on
STMicroelectronics' systems, it only updates the existing infrastructure.

Thanks to Youssef Triki <Youssef.Triki@st.com> for the original patches.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>