[STM,FIX] Enable the configuration of PIO40-42 on STxH407.
The function stxh407_pioalt_pad() did not deal with configuration of
the 3 PIO ports PIO40-42 in the FLASH Bank well - it just called BUG()!
This commit deals with the configuration of these 3 PIO ports better.
Specifically, it configures the open-drain (OD) correctly, and it now
deals with the output-enable (OE) and pull-up (PU) by doing nothing!
Specifically, it no longer calls BUG()!

Being pedantic, we should configure the PU and PD features in the "FLASH
Bank" explicitly instead using the four SYS_CFG registers: 3080, 3081,
3085, and 3086. However, this would necessitate passing in the alternate
function number to the function stxh407_pioalt_pad(), and adding some
horrible complexity to it.  Alternatively, we could just perform 4 32-bit
"pokes" to these four SYS_CFG registers early in the initialization.
In practice, these four SYS_CFG registers are correct after a reset,
and U-Boot does not need to change them, so we (cheat and) rely on these
registers being correct.
WARNING: Please be aware of this (pragmatic) behaviour!

Empirically, this commit allows eMMC/SD to work correctly on a B2120.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>