[MMC] Allow MMC boot-mode to use 1-bit (or 4-bits).
The MMC's boot-mode controller can operate in 1-bit or 4-bit boot-mode.
Prior to this commit, U-Boot implicitly had an assumption that the
boot-mode controller should (unconditionally) always operate in 4-bit
mode. The code explicitly enforced this implicit assumption.  However the
MMC boot-mode controller used by STMicroelectronics (and possibly others)
has a H/W limitation, whereby it can *only* operate in 1-bit boot-mode.

This commit now allows the user to configure if the MMC boot-mode should
be 4-bit (still the default), or just 1-bit. To use 1-bit boot-mode,
the following macro should now be defined:

	#define CONFIG_MMC_BOOT_MODE_1_BIT

This was tested by building U-Boot for the B2120 board, and burning the
resultant "u-boot.bin" image to the first MMC boot partition ("mmc dev 0
1"), and configuring the board to boot directly from the on-board eMMC
flash device. The board booted successfully! To be clear, without this
patch, boot-from-eMMC always failed on this board.

Testing was limited to a single board:

	B2120-A02 (STiH407-HDK-AM00-045) with a STiH407 cut 1.0

NOTE: In order to read/write from/to eMMC/SD devices correctly, it should
be noted that (empirically) the minimum version of the "TargetPack"
required is revision "v6".  Earlier versions of the TargetPacks do
not configure the MMC controller correctly, resulting in errors being
reported. The "*.romgen" file committed in U-Boot (at the time of writing)
was based on "v3" of the TargetPacks, which is too old for correct
MMC operation.  The perennial advice is always to run "make romgen",
and generate a new "*.romgen" file using the *latest* TargetPacks.
For the testing undertaken for this commit, "v8" was used successfully.

This patch was written by Sean McGoogan, but was based on a patch provided
by Youssef Triki <Youssef.Triki@st.com>. Thanks Youssef!

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>