[STM] Perform a dummy read prior to initialising the MMC driver.
When we are booting directly from eMMC, then the STMicroelectronics' SoC
(STxH407) has a requirement that there must be at least one eMMC boot-mode
"transaction", before we can initialize the eMMC driver safely.

Hence, when in boot-from-eMMC mode, we will now perform a single "dummy"
read to the boot-mode region. We simply read from physical address zero,
and discard the result.  This read only needs to be performed once.

It should be noted that this is only really an issue when booting via
GDB/JTAG, and when the mode-pins are in boot-from-eMMC mode, when the
boot-controller has not yet been exercised!

Testing was limited to a single board:

	B2120-A02 (STiH407-HDK-AM00-045) with a STiH407 cut 1.0

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>