[STM] Route the PIO muxing for the FSM SFC (on the STxH407)
Correctly route the PIO muxing in the FlashSS (Flash Sub-System),
specifically for the FSM SPI Serial Flash Controller (SFC), on the
STxH407 SoC from STMicroelectronics.

This is (strictly speaking) not required when booting from SPI, as the
H/W automatically (implicitly) configures this correctly when booting
from SPI – but it is (explicitly) required to be done by S/W, in order
to use the SPI Serial Flash Controller, when *not* booting from SPI.
However, we will unconditionally do this when initializing the SPI FSM
driver (irrespective of the actual boot device).

Testing was limited to a single board, with the FSM driver enabled and
the mode-pins set to boot-from-eMMC, using:

	B2120-A02 (STiH407-HDK-AM00-045) with a STiH407 cut 1.0

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>