[STM] Add new target of "b2120a9_emmc_config"
Added a new top-level configuration make target of "b2120a9_emmc_config",
which will build for a boot-from-eMMC environment on STMicroelectronics'
B2120 board, for an ARM A9 core.

Testing included building for the B2120 board, with *both* boot-from-SPI,
and boot-from-eMMC environments, and burning the resulting "u-boot.bin"
images directly into SPI and eMMC (respectively). The board was
booted multiple times, loading "u-boot.bin" over NFS, re-flashing, and
re-booting. In addition U-Boot environment variables were modified and
saved to both SPI and MMC, and tested by printing them after a reboot.
This included saving the environment in SPI, whilst booting from eMMC
(and vice versa).  All tests were successful, with no problems noted.

In summary, it looks like both SPI and eMMC are now stable and robust,
with no known problems with either type of flash.

Testing was limited to a single board:

	B2120-A02 (STiH407-HDK-AM00-045) with a STiH407 cut 1.0

NOTE: In order to read/write from/to eMMC/SD devices correctly, it should
be noted that (empirically) the minimum version of the "TargetPack"
required is revision "v6".  Earlier versions of the TargetPacks do
not configure the MMC controller correctly, resulting in errors being
reported. The "*.romgen" file committed in U-Boot (at the time of writing)
was based on "v3" of the TargetPacks, which is too old for correct
MMC operation.  The perennial advice is always to run "make romgen",
and generate a new "*.romgen" file using the *latest* TargetPacks.
For the testing undertaken for this commit, "v8" was used successfully.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>