[STM] Allow "savenv" to eMMC/SD on the B2120 board.
Add support to save U-Boot's environment variables to MMC (eMMC/SD)
devices on the B2120 board from STMicroelectronics.

By default on this board, SPI will still be used, if SPI is configured.
If saving the environment to MMC is enabled, then by default it will
be saved to the first enabled MMC device (e.g. MMC(0)), which will by
default be the on-board eMMC device.

Testing was limited to a single board:

	B2120-A02 (STiH407-HDK-AM00-045) with a STiH407 cut 1.0

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>