[STM] Add support for the STxH410 SoC (on the B2120 board).
Add support for the STxH410 SoC on the B2120 board (both from
STMicroelectronics). The STxH410 is very similar to the STxH407, and so
we use the SoC files for the STxH407 SoC, when we use the STxH410 SoC.

Both the STxH407 and STxH410 SoCs can be used on the same B2120 board.

Testing was limited to a single board:

	B2120-A02 (STiH407-HDK-AM00-045)

With 3 different SoCs (in turn):

	STxH410 cut 1.0		(sample #963)
	STxH407 cut 1.0		(sample #122)
	STxH407 cut 3.0

Testing included booting directly from both SPI and eMMC. No problems
were noted.  The "bdinfo" command also correctly identifies if the SoC
is actually a STxH407 or a STxH410 (and its major cut number).

When using the STxH407, then "make romgen" was first re-run using
TargetPack "v8".  As always, the perennial advice is to run "make romgen",
and generate a new "*.romgen" file using the *latest* TargetPacks.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>
(cherry picked from commit 1a02c8b25b0e95c53dc116bc299b88331aaf0df7)

Conflicts:

	Makefile
	common/cmd_bdinfo.c

Signed-off-by: Sachin Verma <sachin.verma@st.com>