[STM] Add support for the Macronix MX25L25635F SPI Serial Flash device
Add support for the Macronix MX25L25635F SPI Serial Flash device, in
the STMicroelectronics tree.

In addition, support for the full 32 MiB is enabled, as 32-bit addressing
(using 4-byte commands) is now present, as this is supported in the "F"
revision of this SPI device (c.f. the "E" revision).

NOTE: the MX25L25635E and the MX25L25635F report having the exact same
Device ID, but the MX25L25635E does *not* support the 4-byte addressing
instructions.

Tested whilst adding support for the B2089 board:

	STiH407-HVK (B2089A-00-E028), with a STxH407 cut 2.x SoC

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>