[STM] Add support for the B2089 board
This commit adds support for the B2089 board from STMicroelectronics,
which may be fitted with either a STxH407 or a STxH410 SoC. This board is
"similar" to the B2120 board, and hence the board-specific files have
also been placed in the "board/st/b2120" directory, to maximise re-use,
and to minimise maintenance.

This includes support for the new improved "FlashSS" IP block, including
the following Flash types:

	* SPI Serial Flash devices
	* raw SLC NAND devices	(B2006/B2007)	[Hamming + BCH]
	* eMMC devices		(B2143)
	* SD cards		(B2144)

It should be noted that the B2089 only has SPI Serial Flash on-board, and
*all* the other flashes require a suitable daughter board to be fitted
(such as the B2006, B2007, B2143, B2144). In particular, boot-from-NAND
and boot-from-eMMC is only possible when a suitable daughter board is
fitted. Note, boot-from-SD cards is *not* possible.

Testing was limited to a single board:

	STiH407-HVK (B2089A-00-E028), with a STxH407 cut 2.x SoC

In addition, testing used the following Flash daughter boards:

	B2007A-02-N052	with a 2KiB page NAND device (BCH18)
	B2143A-R27	with a eMMC device (+mod to short CD to GND)
	B2144A-00-R029	with a few SD cards

Testing included booting directly from all 3 possible bootable flash
devices: SPI, NAND and eMMC - and copying u-boot.bin over NFS, in turn. In
the case of NAND and eMMC, it was using the appropriate daughter boards
listed above, in the case of SPI Serial Flash, it was using the on-board
device.  No problems were noted.

In addition various other commands/peripherals were tested. Of note,
only "networking" need be mentioned. Although both "ping" and "nfs"
worked (on all 3 RJ-45 ports), there were numerous time-outs with NFS,
and performance appeared to be below standard. It is unclear if this is
a SoC or board or S/W issue! Additional testing should be performed with
NFS to be sure, on other boards/SoCs.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>