[STM] Migrated ST boards to boards.cfg method
This migration to new boards.cfg method of configuration is a step
towards preparing for rebase to v2014.01 of uboot. Most of the other
platforms have already moved to this method of configuration.

         This method precludes the need to maintain the Makefile for
different board configurations. Configurations are listed in a rather
neat tabular format. ST platforms were making use of config.tmp file
to override CONFIG_SYS_TEXT_BASE for different boards. After this
migration there is no need to maintain this config.tmp file. As a
consequence, there is no need to include this config.tmp file in the
board specific *config.mk* files. I have removed those entries.

         Another change is the way romgen tool is invoked. I have
added the romgen tool parameters as part of board options in the
board.cfg file. This way, it is inline with the u-boot philosophy
of delineating configuration. To cover this functionality a new
configuration variable CONFIG_ROMGEN_PARAMS is now added. Addition
of any new board should initialize this variable appropriately with
parameters as expected by *mkromgen* script in the u-boot tree at
root.
         ROMGEN_PARAMS contains the parameters fed to mkromgen script
but with a slight change. you need to replace all *spaces* with '@'
character and all *commas* with '#' character. This change was done in
order to keep the mkconfig script happy. Choice of these two characters
is entirely ours. In this case, i have chosen '@' and '#'.

        From a user's perspective the sequence of configuration remains
same apart from addition of a new step. For instance, sequence for
b2020 board booting through SPI is:
        1. make distclean
        2. make b2020a9_spi_config
        3. make romgen_prepare
        4. make romgen
        5. make

Step 3 is the additional step now. Steps 3 and 4 are required only if
you are not happy with the supplied romgen files and wish to tweak some-
things.

Note: I have tested this new migration only on B2020 (sthx416) and
B2120 (sthx407) for booting u-boot image through gdb.

Signed-off-by: Sachin Verma <sachin.verma@st.com>