[STM] Move STM global field refs to arch_global_data
Following on from the recent (incomplete) "snapshot" of the merge of
the official 'v2014.01' release, into the STMicroelectronics' git tree,
this commit is one step (of many steps) towards the final completion of
that aforementioned snapshot commit.

The main-line tree had moved the architecture specific global data
fields from structure global_data to a new architecture specific
structure called arch_global_data.

This commit updates all references of such global fields which are
specific to ST's ARM based platforms to new data structure.

It should be noted that this tree still does *not* build (on either ST40
or ARM), and further work still needs to be done to get it to build.

Also, corresponding changes are still *NOT* done for ST40 architecture.
ST40 changes would be done in later commits.

Signed-off-by: Sachin Verma <sachin.verma@st.com>