[STM] Re-comment "bit-banging" to "EMI bit-banging" for NAND.
For STMicroelectronics' SoCs, when using NAND, as well as there being
drivers to support the H/W (Flex-Mode) controller, there was also a
"bit-banging" driver which was usually described as a "S/W bit-banging"
driver. This driver was used when the following configuration was defined:

		CONFIG_SYS_STM_NAND_USE_BIT_BANGING

In hindsight, (and being pedantic) the appellation of "S/W bit-banging"
is slightly misleading, as it was never a *pure* S/W driver, as it did
use the (parallel-NOR) EMI H/W, and so was more of a "hybrid" driver!

With the release of the latest SoCs from STMicroelectronics which now
use the (new) "FlashSS" (Flash Sub-system) IP block, and the removal of
the legacy EMI H/W, then things have changed somewhat. Specifically:

	1) It is no longer possible to to use the "EMI" with NAND to
	   effect a pseudo-bit-banging driver on the latest SoCs.

	2) It is now possible to use the FlashSS IP block with a pure
	   S/W bit-banging driver, by re-configuring its pins as GPIOs!

It is not obvious that there is any real advantage in using the FlashSS
to implement a pure S/W bit-banging driver, but possible it now is!

So, to entertain this possibility, and to reduce any possible confusion
in the future, this commit globally updates the source comments to try
and be clearer on the nomenclature of this "hybrid" driver.

Specifically, all previous references to "S/W bit-banging" (in a NAND
context only), will now say "EMI bit-banging" instead, to make it
clearer that the EMI is an integral part of this NAND driver. Hence,
use of the term "S/W bit-banging" can now freely be used (with impunity)
to add support for a pure S/W bit-banging NAND driver using the raw GPIO
pins in the FlashSS IP block (if ever required).

Finally, it should be noted that the above configuration name has *not*
been modified. Hence, this commit make no logical changes to the code -
it is purely just a big exercise in clarification to lots of comments!

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>