[STM] Prepare to add support for the FlashSS
Prior to the creation of the (new) FlashSS (Flash Sub-System) IP block,
various FLASH-specific configuration registers were stored in a H/W
block known as the "EMISS" (aka the "EMIPCISS"). With the advent of the
FlashSS these registers have now all been "re-homed", and instead exist
in the FlashSS IP block, on SoCs where it is present.

So, ideally, these register definitions should now all be renamed to show
the appropriate containing IP block for them. Making it clear if they
are in the (older) EMISS or the (newer) FlashSS block. However, some of
the code does not care which IP block they are in, but mealy wishes to
use an macro to refer to them.  Hence, we now need 3 distinct name-spaces:

	STM_EMISS_*	For legacy EMISS/EMIPCISS registers/block.
	STM_FLASHSS_*	For the new FlashSS registers/block.
	STM_FLASH_*	IP-block agnostic (i.e. abstracted).

This commit is one step toward this re-factoring goal, by creating the new
"abstracted" macro names, and renaming most of the old macros such that:

	STM_EMISS_* -> STM_FLASH_*

It should be noted, that the only remaining macro with the prefix
STM_EMISS_* should now only be "STM_EMISS_REGS_BASE", as this is still
the definition of the base-address for the EMISS/PMIPCISS block on SoCs
that use that H/W block. That macro should NOT be defined for SoCs using
the newer FlashSS.

In addition, for the same rationale, the function emiss_nandi_select(),
has also been renamed to stm_flash_nandi_select(), to remove the (now
inappropriate) "emiss" prefix it had.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>
(cherry picked from commit dd29698738cd8ae9ea4364ef6b70513a8420b787)