[STM] Add proper support for FlashSS on the STxH407/10.
This commit completes the support for the newer FlashSS IP block, that
is replacing the (legacy) EMISS/EMIPCISS block, in the latest generation
of SoCs from STMicroelectronics.

This commit adds FlashSS support to the STxH407/STxH410 SoCs, which
was specifically required to add support for NAND (in the FlashSS)
on those SoCs.

In addition, this commit also moves the definitions for the base addresses
(for SPI and MMC) from a board-specific file, to a SoC-specific file,
so there will be fewer definitions, and better re-use!

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>