[STM] Update Makefiles to new Kconfig style
Following on from the recent (incomplete) "snapshot" of the merge of
the official 'v2014.01' release, into the STMicroelectronics' git tree,
this commit is one step (of many steps) towards the final completion of
that aforementioned snapshot commit.

U-Boot mainline has moved to a new style of Makefiles. U-Boot now
uses Kconfig style of Makefiles. This patch migrates all the existing
STMicroelectronics specific makefiles to this new style.

It should be noted that this tree still does *not* build (on either ST40
or ARM), and further work still needs to be done to get it to build.

Signed-off-by: Sachin Verma <sachin.verma@st.com>