[STM] Removing unused remnants of CONFIG_I2C_CMD_TREE
Following on from the recent (incomplete) "snapshot" of the merge of
the official 'v2014.01' release, into the STMicroelectronics' git tree,
this commit is one step (of many steps) towards the final completion of
that aforementioned snapshot commit.

CONFIG_I2C_CMD_TREE is no longer needed as flat i2c commands are no
longer available in the U-Boot sources.

It should be noted that this tree still does *not* build (on either ST40
or ARM), and further work still needs to be done to get it to build.

Signed-off-by: Sachin Verma <sachin.verma@st.com>