[STM] Miscellaneous Fixes after rebase to v2014.01
Following on from the recent (incomplete) "snapshot" of the merge of
the official 'v2014.01' release, into the STMicroelectronics' git tree,
this commit is one step (of many steps) towards the final completion of
that aforementioned snapshot commit.

Miscellaneous fixes to get nand, i2c, ethernet, bootm in a consistent
state w.r.t. the new U-Boot code base (v2014.01). This patch tries to
address the remaining compilation issues.

It should be noted that this tree still does *not* build (on either ST40
or ARM), and further work still needs to be done to get it to build.

Signed-off-by: Sachin Verma <sachin.verma@st.com>