[STM] Migrate ST boards to boards.cfg v2014.01 style
Following on from the recent (incomplete) "snapshot" of the merge of
the official 'v2014.01' release, into the STMicroelectronics' git tree,
this commit completes that aforementioned snapshot commit.

This commit aligns configuration of ST boards according to the
re-ordering of fields, plus addition of some new fields.

Previous step of aligning with boards.cfg was done in this
commit: 2ef0df6adfdcfb3da1a5ed329ffac880ea2b9795

  From a user's perspective the sequence of configuration would be
as follows (for b2120 board):
            1. make distclean
            2. make b2120a9_spi_config
            3. make romgen_prepare
            4. make romgen
            5. make

Steps 3 and 4 are required only if you are not happy with the supplied
romgen files and wish to tweak some-things.

This commit completes rebase to v2014.01 of U-Boot.

This tree compiles and build for ARM.
I have tested it to boot using gdb on B2120 (STih407) board.

Signed-off-by: Sachin Verma <sachin.verma@st.com>