Enabled D-cache support for B2120.
Ethernet driver needs to be modified to avoid the issues caused byin-coherent
caches and stale data in RAM. DMA descriptors are shared between CPU and the DMA
and are modified by each of them to indicate current status and current owner of a
particular descriptor.

In situations where D-cache is enabled CPU or DMA may get wrong
status of the descriptor This patch avoids such races along with other chnages in ethernet
driver.